[FIX] Moved csr files in adequacy_patch_csr/
