+++ src/solver/cmake/solver.cmake
@@ -59,8 +59,8 @@ set(RTESOLVER_OPT
 		optimisation/weekly_optimization.cpp
 		optimisation/adq_patch_local_matching.h
 		optimisation/adq_patch_local_matching.cpp
-		optimisation/adq_patch_curtailment_sharing.h
-		optimisation/adq_patch_curtailment_sharing.cpp
+		optimisation/adequacy_patch_csr/adq_patch_curtailment_sharing.h
+		optimisation/adequacy_patch_csr/adq_patch_curtailment_sharing.cpp
 		optimisation/adequacy_patch_csr/solve_problem.cpp
 		optimisation/adequacy_patch_csr/set_variable_boundaries.cpp
 		optimisation/adequacy_patch_csr/set_problem_cost_function.cpp
