+++ src/solver/optimisation/adequacy_patch_csr/adq_patch_curtailment_sharing.h
