+++ src/tests/src/solver/optimisation/adequacy_patch.cpp
@@ -5,7 +5,7 @@
 #include <boost/test/included/unit_test.hpp>
 
 #include "adq_patch_local_matching.h"
-#include "adq_patch_curtailment_sharing.h"
+#include "adequacy_patch_csr/adq_patch_curtailment_sharing.h"
 
 #include <vector>
 #include <tuple>
